added scheduler type selection and answer tokens dtype fixed

+++ transactions_qa/tasks/context_amnt_tasks.py
@@ -146,7 +146,7 @@ class MeanAmountBinnedTaskBinary(NumericTaskAbstract):
         batch_answer_template_encoded = answer_template_encoded.repeat(batch_size, 1)
         # Answer template encoding + target tokens + EOS token
         batch_answer_encoded = torch.cat([batch_answer_template_encoded,
-                                          target_encoded_batch['input_ids']], dim=1).to(device)
+                                          target_encoded_batch['input_ids']], dim=1).long().to(device)
         # Answer masks
         batch_answer_template_mask = torch.ones(batch_size, answer_template_encoded.shape[1]).to(device)
         batch_answer_mask = torch.cat([batch_answer_template_mask,
@@ -394,7 +394,7 @@ class MeanAmountNumericTaskBinary(NumericTaskAbstract):
         batch_answer_template_encoded = answer_template_encoded.repeat(batch_size, 1)
         # Answer template encoding + target tokens + EOS token
         batch_answer_encoded = torch.cat([batch_answer_template_encoded,
-                                          target_encoded_batch['input_ids']], dim=1).to(device)
+                                          target_encoded_batch['input_ids']], dim=1).long().to(device)
         # Answer masks
         batch_answer_template_mask = torch.ones(batch_size, answer_template_encoded.shape[1]).to(device)
         batch_answer_mask = torch.cat([batch_answer_template_mask,
@@ -861,7 +861,7 @@ class MeanAmountNumericTaskOpenEnded(NumericTaskAbstract):
         batch_answer_template_encoded = answer_template_encoded.repeat(batch_size, 1)
         # Answer template encoding + target tokens + EOS token
         batch_answer_encoded = torch.cat([batch_answer_template_encoded,
-                                          target_encoded_batch['input_ids']], dim=1).to(device)
+                                          target_encoded_batch['input_ids']], dim=1).long().to(device)
         # Answer masks
         batch_answer_template_mask = torch.ones(batch_size, answer_template_encoded.shape[1]).to(device)
         batch_answer_mask = torch.cat([batch_answer_template_mask,
